Reap Timer + Actually use r.config
We were just parsing the user list straight from a new call to /robin before. Now we'll use the copy of r.config the page is passed and update it after 60s.

The timer is just pulled from the server and then updated based on the local time. There's an opportunity to
1) See if it drifts by comparing it to the versions in the new copies of r.config we pull down
2) Account for difference between r.config.server_time and local time